Finish splitting patch maker tests out by architecture
